Connection to AWS' database and models migration. default apps and apps created are in database and it works correctly :)

+++ Mangcoin/apps/tokens/models.py
@@ -4,4 +4,4 @@ class tokens(models.Model):
     token_id = models.CharField(max_length=10, primary_key=True)
     token_name = models.CharField(max_length=255)
     fundation_year = models.PositiveSmallIntegerField(blank=True)
-    total_aumont = models.PositiveSmallIntegerField(blank=True)
+    total_aumont = models.PositiveBigIntegerField(blank=True)
